+++ api/models/response.js
@@ -23,16 +23,19 @@ const responseSchema = new Schema({
   },
   quiz: Schema.Types.Mixed,
   textelicitation_drugOverdose: Schema.Types.Mixed,
-  quiz2: Schema.Types.Mixed,
-  quiz3: Schema.Types.Mixed,
+  attitude_Elicitation: Schema.Types.Mixed,
+  topic_Involvement: Schema.Types.Mixed,
   textelicitation_AmericanPopulation: Schema.Types.Mixed,
   textelicitation_Opioids: Schema.Types.Mixed,
   noelicitation_drugOverdose: Schema.Types.Mixed,
   noelicitation_AmericanPopulation: Schema.Types.Mixed,
-  quiz8: Schema.Types.Mixed,
-  viz: Schema.Types.Mixed,
+  noelicitation_Opioids: Schema.Types.Mixed,
+  recall_drugOverdose: Schema.Types.Mixed,
   viz1: Schema.Types.Mixed,
-  viz2: Schema.Types.Mixed,
+  visualElicitation_drugOverdose:Schema.Types.Mixed,
+  visualElicitation_Opioids:Schema.Types.Mixed,
+  visualElicitation_population:Schema.Types.Mixed,
+  recall_Opioids: Schema.Types.Mixed,
   cogref: Schema.Types.Mixed,
   cogref1: Schema.Types.Mixed,
   instructions: Schema.Types.Mixed,
