+++ client/src/pages/pageHandler.js
@@ -3,19 +3,22 @@ const pages = [
   "pre",
   "instructions1",
   "quiz",
-  "quiz2",
-  "quiz3",
+  "attitude_Elicitation",
+  "topic_Involvement",
   "textelicitation_drugOverdose",
   "textelicitation_AmericanPopulation",
   "textelicitation_Opioids",
   "noelicitation_drugOverdose",
   "noelicitation_AmericanPopulation",
-  "quiz8",
+  "noelicitation_Opioids",
+  "visualElicitation_drugOverdose",
+  "visualElicitation_population",
+  "visualElicitation_Opioids",
   "cogref",
   "cogref1",
   "viz1",
-  "viz",
-  "viz2",
+  "recall_drugOverdose",
+  "recall_Opioids",
   "debrief",
 ];
