+++ client/src/pages/survey/attitude_Elicitation.js
@@ -1,4 +1,5 @@
 import React, { useRef } from "react";
+import { atom, selector } from "recoil";
 import { useHistory, useLocation } from "react-router-dom";
 import pageHandler from "../pageHandler";
 import axios from "axios";
@@ -10,7 +11,7 @@ import { useRecoilValue } from "recoil";
 import { questionState } from "../../atoms/questionSelector";
 import "survey-react/survey.css";
 
-const PreSurveyPage1 = (props) => {
+const Attitude_Elicitation = (props) => {
   const quizResponses = useRef([]);
   const history = useHistory();
   const location = useLocation();
@@ -19,67 +20,67 @@ const PreSurveyPage1 = (props) => {
   const extraQuestions =
     questionCondition == "strength"
       ? [
-          
+       
         ]
       : [];
 
+
+      
+
   const json = {
     pages: [
+   
       {
         elements: [
           {
             type: "html",
-            html: "<p style='font-size: 22px;'>Since 2002, the number of Americans who have died every year from overdoses of <span style='font-weight: bold;'>synthetic opioids...</span>  </p>",
-           
+            html: "<h4><h4/>",
           },
-            {
-              name: "",
-              type: "radiogroup",
-              title: "",
-              // isRequired: true,
-              // choices: ["yes", "no"],
-              },
-              {
-                type: "html",
-                name: "image_and_text",
-                html: "<div style='text-align: center'><img src='https://raw.githubusercontent.com/subham27-07/youdrawitnew/main/d.JPG' width='80%' height='100%' /><br/><br/><p style='text-align: justify'>...has increased by more than <span style='font-weight: bold;'> 5451 percent </span>. Substance use disorders refers to direct deaths from overdoses of illicit drugs synthetic opioids (mostly fentanyl).  “We know that substance use is more dangerous than it has ever been, as fentanyl has continued to permeate the illicit drug supply, increasing the risk for overdoses among both people with substance use disorders as well as those who use drugs occasionally,” said Dr. Nora Volkow, director of the National Institute on Drug Abuse. Deaths involving synthetic opioids such as fentanyl increased by a marked 18% in 2021, according to the CDC data. Deaths involving cocaine and psychostimulants such as methamphetamine were also significantly more frequent, while those involving heroin decreased.</p></div>",
-            },
-          
+          {
+            name: "claim",
+            type: "radiogroup",
+            title: ` "What is your opinion on drug overdose in US ?"`,
+            isRequired: true,
+            choices: [
+                "Extremely serious problem",
+                "serious problem",
+                "Moderate problem",
+                "Minor Problem",
+                "Not at all a problem",
+            ],
+            // correctAnswer: "a conclusion about a topic",
+          },
+          {
+            name: "new",
+            type: "radiogroup",
+            title: ` "Should the US make combating drug abuse and overdose a priority, i:e allocating tax dollars to treatment and prevention programs?" `,
+            isRequired: true,
+            choices: [
+                "High Priority",
+                "Moderate Priority",
+                "Neutral",
+                "Low Priority",
+                "Not a Priority"
+            ],
+            // correctAnswer: "a news headline",
+          },
+          {
+            name: "headline",
+            type: "radiogroup",
+            title: ` "What is your opinion on drug legalization and decrimination in the US?" `,
+            isRequired: true,
+            choices: [
+                "Strongly Oppose",
+              "Somewhat Oppose",
+              "Neutral",
+              "Somewhat Favor",
+              "Strongly Favor",
+            ],
+            // correctAnswer: "a news headline",
+          },
+          ...extraQuestions,
         ],
       },
-      // {
-      //   elements: [
-      //     {
-      //       type: "html",
-      //       html: "<h4>We are asking you to respond to these questions to make sure you understand the task at hand. You will not be able to move forward if you answer incorrectly.<h4/>",
-      //     },
-      //     {
-      //       name: "claim",
-      //       type: "radiogroup",
-      //       title: `The tweet: "Spielberg is one of the worst directors of the recent decade." is ___.`,
-      //       isRequired: true,
-      //       choices: [
-      //         "a conclusion about a topic",
-      //         "a news headline",
-      //         "I don't know",
-      //       ],
-      //       // correctAnswer: "a conclusion about a topic",
-      //     },
-      //     {
-      //       name: "headline",
-      //       type: "radiogroup",
-      //       title: `The tweet: "Steven Spielberg's latest three movies were among the worst rated in Rotten Tomatoes." is ___.`,
-      //       isRequired: true,
-      //       choices: [
-      //         "a conclusion about a topic",
-      //         "a news headline",
-      //         "I don't know",
-      //       ],
-      //       // correctAnswer: "a news headline",
-      //     },
-      //     ...extraQuestions,
-      //   ],
-      // },
     ],
   };
 
@@ -120,7 +121,7 @@ const PreSurveyPage1 = (props) => {
     // console.log(options);
 
     console.log("Survey results: " + JSON.stringify(quizResponses.current));
-    axios.post("/api/quiz8", quizResponses.current).then((response) => {
+    axios.post("/api/attitude_Elicitation", quizResponses.current).then((response) => {
       let nextPage = pageHandler(location.pathname);
       history.push(nextPage);
     });
@@ -128,7 +129,7 @@ const PreSurveyPage1 = (props) => {
 
   const onCurrentPageChanging = (survey, option) => {
     if (!option.isNextPage) return;
-    let allTrue = false;
+    let allTrue = true;
     survey.getAllQuestions().forEach((q) => {
       if (survey.currentPage == q.page) {
         let correct = isAnswerCorrect(q);
@@ -139,11 +140,11 @@ const PreSurveyPage1 = (props) => {
       }
     });
     console.log(allTrue);
-    // if (allTrue) {
-    //   option.allowChanging = true;
-    // } else {
-    //   option.allowChanging = false;
-    // }
+    if (allTrue) {
+      option.allowChanging = true;
+    } else {
+      option.allowChanging = false;
+    }
     // console.log(survey.currentPage());
     // option.oldCurrentPage.questions.forEach((q) => {
     //   console.log(q);
@@ -189,8 +190,6 @@ const PreSurveyPage1 = (props) => {
 
   const model = new Survey.Model(json);
   model.showCompletedPage = false;
-  model.questionTitleTemplate = "";
-  model.showQuestionNumbers = "none";
   model.onTextMarkdown.add((sender, options) => {
     var text = options.text;
     var html = getTextHtml(text, correctStr, true);
@@ -221,8 +220,8 @@ const PreSurveyPage1 = (props) => {
           justifyContent: "center",
         }}
       >
-        <Typography variant="h3">
-          How Bad Is the   <span style={{ fontWeight: "bold" }}> Drug Overdose... </span> Epidemic?   
+        <Typography variant="h5">
+        Please Answer the questions below👇.
         </Typography>
         <Divider></Divider>
         
@@ -238,4 +237,31 @@ const PreSurveyPage1 = (props) => {
   );
 };
 
-export default PreSurveyPage1;
+export const labelSelector = selector({
+    key: "labelQuestionSelector",
+    get: ({ get }) => {
+      let questionCondition = get(questionState);
+      switch (questionCondition) {
+        case "share":
+          return [
+            "Definitely no",
+            "Probably no",
+            "Probably yes",
+            "Definitely yes",
+          ];
+          break;
+        case "strength":
+          return [
+            "Extremely Serious Problem",
+            "Serious Problem",
+            "Moderate Problem",
+            "Minor Problem",
+            "Not a Problem",
+          ];
+  
+          break;
+      }
+    },
+  });
+
+export default Attitude_Elicitation;
