+++ client/src/pages/survey/noelicitation_Opioids.js
@@ -1,139 +1,16 @@
-import React, { useRef, lazy, Suspense } from 'react';
-
-
+import React, { useRef } from "react";
 import { useHistory, useLocation } from "react-router-dom";
 import pageHandler from "../pageHandler";
 import axios from "axios";
 import * as Survey from "survey-react";
-import { Divider, Typography, Container, Button } from "@mui/material";
+import { Divider, Typography, Container } from "@mui/material";
 import Tweet from "../../components/tweet/tweet";
 import TweetQuote from "../../components/tweet/tweetQuote";
 import { useRecoilValue } from "recoil";
 import { questionState } from "../../atoms/questionSelector";
 import "survey-react/survey.css";
-const LineChart = lazy(() => import('./visualizations/LineChart3'));
 
-
-const Viz2 = () => {
-  const lineData = [
-    {
-      id: 1,
-      value: 730,
-      year: 1999,
-    },
-    {
-      id: 1,
-      value: 782,
-      year: 2000,
-    },
-    {
-      id: 1,
-      value: 957,
-      year: 2001,
-    },
-    {
-      id: 1,
-      value: 1295,
-      year: 2002,
-    },
-    {
-      id: 1,
-      value: 1400,
-      year: 2003,
-    },
-    {
-      id: 1,
-      value: 1664,
-      year: 2004,
-    },
-    {
-      id: 1,
-      value: 1742,
-      year: 2005,
-    },
-    {
-      id: 1,
-      value: 2707,
-      year: 2006,
-    },
-    {
-      id: 1,
-      value: 2213,
-      year: 2007,
-    },
-    {
-      id: 1,
-      value: 2306,
-      year: 2008,
-    },
-    {
-      id: 1,
-      value: 2946,
-      year: 2009,
-    },
-    {
-      id: 1,
-      value: 3007,
-      year: 2010,
-    },
-    {
-      id: 1,
-      value: 2666,
-      year: 2011,
-    },
-    {
-      id: 1,
-      value: 2628,
-      year: 2012,
-    },
-    {
-      id: 1,
-      value: 3105,
-      year: 2013,
-    },
-    {
-      id: 1,
-      value: 5544,
-      year: 2014,
-    },
-    {
-      id: 1,
-      value: 9580,
-      year: 2015,
-    },
-    {
-      id: 1,
-      value: 19413,
-      year: 2016,
-    },
-    {
-      id: 1,
-      value: 28466,
-      year: 2017,
-    },
-    {
-      id: 1,
-      value: 31335,
-      year: 2018,
-    },
-    {
-      id: 1,
-      value: 36359,
-      year: 2019,
-    },
-    {
-        id: 1,
-        value: 56516,
-        year: 2020,
-    },
-    {
-        id: 1,
-        value: 70601,
-        year: 2021,
-    },
-  ];
-
-
+const Noelicitation_Opioids = (props) => {
   const quizResponses = useRef([]);
   const history = useHistory();
   const location = useLocation();
@@ -142,66 +19,67 @@ const Viz2 = () => {
   const extraQuestions =
     questionCondition == "strength"
       ? [
-        
+          
         ]
       : [];
 
   const json = {
     pages: [
-    
       {
         elements: [
-        {
+          {
+            type: "html",
+            html: "<p style='font-size: 22px;'>Since 2002, the number of Americans who have died every year from overdoses of <span style='font-weight: bold;'>synthetic opioids...</span>  </p>",
+           
+          },
+            {
+              name: "",
+              type: "radiogroup",
+              title: "",
+              // isRequired: true,
+              // choices: ["yes", "no"],
+              },
+              {
                 type: "html",
-                html: "<p style='font-size: 22px;'>Since 2002, the number of Americans who have died every year from overdoses of <span style='font-weight: bold;'>synthetic opioids...</span>  </p>",
-               
-        },
-          {
-            name: "claim",
-            type: "radiogroup",
-            title: ` "I would recommend this article to my family and friends"`,
-            isRequired: true,
-            choices: [
-                "Not at All",
-                "A little",
-                "Moderately",
-                "A lot",
-                "Extremely",
-            ],
-            // correctAnswer: "a conclusion about a topic",
-          },
+                name: "image_and_text",
+                html: "<div style='text-align: center'><img src='https://raw.githubusercontent.com/subham27-07/youdrawitnew/main/d.JPG' width='80%' height='100%' /><br/><br/><p style='text-align: justify'>...has increased by more than <span style='font-weight: bold;'> 5451 percent </span>. Substance use disorders refers to direct deaths from overdoses of illicit drugs synthetic opioids (mostly fentanyl).  “We know that substance use is more dangerous than it has ever been, as fentanyl has continued to permeate the illicit drug supply, increasing the risk for overdoses among both people with substance use disorders as well as those who use drugs occasionally,” said Dr. Nora Volkow, director of the National Institute on Drug Abuse. Deaths involving synthetic opioids such as fentanyl increased by a marked 18% in 2021, according to the CDC data. Deaths involving cocaine and psychostimulants such as methamphetamine were also significantly more frequent, while those involving heroin decreased.</p></div>",
+            },
           
-          {
-            name: "suport",
-            type: "radiogroup",
-            title: ` "The content of this article is surprising to me" `,
-            isRequired: true,
-            choices: [
-                "Not at All",
-                "A little",
-                "Moderately",
-                "A lot",
-                "Extremely",
-            ],
-            // correctAnswer: "a news headline",
-          },
-          {
-            name: "viewOpinion",
-            type: "radiogroup",
-            title: ` "I felt interested in reading this article" `,
-            isRequired: true,
-            choices: [
-                "Not at All",
-                "A little",
-                "Moderately",
-                "A lot",
-                "Extremely",
-            ],
-            // correctAnswer: "a news headline",
-          },
-          ...extraQuestions,
         ],
       },
+      // {
+      //   elements: [
+      //     {
+      //       type: "html",
+      //       html: "<h4>We are asking you to respond to these questions to make sure you understand the task at hand. You will not be able to move forward if you answer incorrectly.<h4/>",
+      //     },
+      //     {
+      //       name: "claim",
+      //       type: "radiogroup",
+      //       title: `The tweet: "Spielberg is one of the worst directors of the recent decade." is ___.`,
+      //       isRequired: true,
+      //       choices: [
+      //         "a conclusion about a topic",
+      //         "a news headline",
+      //         "I don't know",
+      //       ],
+      //       // correctAnswer: "a conclusion about a topic",
+      //     },
+      //     {
+      //       name: "headline",
+      //       type: "radiogroup",
+      //       title: `The tweet: "Steven Spielberg's latest three movies were among the worst rated in Rotten Tomatoes." is ___.`,
+      //       isRequired: true,
+      //       choices: [
+      //         "a conclusion about a topic",
+      //         "a news headline",
+      //         "I don't know",
+      //       ],
+      //       // correctAnswer: "a news headline",
+      //     },
+      //     ...extraQuestions,
+      //   ],
+      // },
     ],
   };
 
@@ -242,7 +120,7 @@ const Viz2 = () => {
     // console.log(options);
 
     console.log("Survey results: " + JSON.stringify(quizResponses.current));
-    axios.post("/api/quiz3", quizResponses.current).then((response) => {
+    axios.post("/api/noelicitation_Opioids", quizResponses.current).then((response) => {
       let nextPage = pageHandler(location.pathname);
       history.push(nextPage);
     });
@@ -250,7 +128,7 @@ const Viz2 = () => {
 
   const onCurrentPageChanging = (survey, option) => {
     if (!option.isNextPage) return;
-    let allTrue = true;
+    let allTrue = false;
     survey.getAllQuestions().forEach((q) => {
       if (survey.currentPage == q.page) {
         let correct = isAnswerCorrect(q);
@@ -261,11 +139,11 @@ const Viz2 = () => {
       }
     });
     console.log(allTrue);
-    if (allTrue) {
-      option.allowChanging = true;
-    } else {
-      option.allowChanging = false;
-    }
+    // if (allTrue) {
+    //   option.allowChanging = true;
+    // } else {
+    //   option.allowChanging = false;
+    // }
     // console.log(survey.currentPage());
     // option.oldCurrentPage.questions.forEach((q) => {
     //   console.log(q);
@@ -311,6 +189,8 @@ const Viz2 = () => {
 
   const model = new Survey.Model(json);
   model.showCompletedPage = false;
+  model.questionTitleTemplate = "";
+  model.showQuestionNumbers = "none";
   model.onTextMarkdown.add((sender, options) => {
     var text = options.text;
     var html = getTextHtml(text, correctStr, true);
@@ -321,15 +201,14 @@ const Viz2 = () => {
       options.html = html;
     }
   });
-  
+
   return (
-    
     <Container
       maxWidth={false}
       style={{
         width: "100%",
         overflow: "auto",
-        minHeight: "600px",
+        height: "100%",
         paddingTop: "30px",
         paddingBottm: "30px",
       }}
@@ -343,26 +222,20 @@ const Viz2 = () => {
         }}
       >
         <Typography variant="h3">
-          How Bad Is the <span style={{ fontWeight: "bold" }}>Drug Overdose</span> Epidemic?
+          How Bad Is the   <span style={{ fontWeight: "bold" }}> Drug Overdose... </span> Epidemic?   
         </Typography>
+        <Divider></Divider>
         
       </div>
-
-      <div className="viz" style={{ display: "flex", flexDirection: "column" }}>
-        <div style={{ width: "100%", height: "500px" }}>
-          <Suspense fallback={<div>loading...</div>}>
-            <LineChart type="value" data={lineData} idLine={1} startYear={2002} />
-          </Suspense>
-        </div>
-        <Survey.Survey
-          model={model}
-          onComplete={onComplete}
-          onCompleting={onCompleting}
-          onCurrentPageChanging={onCurrentPageChanging}
-        />
-      </div>
+      <Divider></Divider>
+      <Survey.Survey
+        model={model}
+        onComplete={onComplete}
+        onCompleting={onCompleting}
+        onCurrentPageChanging={onCurrentPageChanging}
+      />
     </Container>
   );
-}
+};
 
-export default Viz2;
+export default Noelicitation_Opioids;
