+++ client/src/pages/survey/visualizations/LineChart2.js
@@ -1,8 +1,6 @@
 import React, { Component } from 'react';
 import * as d3 from 'd3';
 
-
-
 const marginConvention = (selection, props) => {
   const {
     width, height, margin, className = 'margin-group',
