+++ client/src/pages/survey/visualizations/LineChartTest.js
@@ -63,17 +63,16 @@ class LineChart extends Component {
    
     const x = d3.scaleLinear().range([0, innerWidth]);
     const y = d3.scaleLinear().range([innerHeight, 0]);
-    y.domain([0, d3.max(data, d => d[type])]);
+    y.domain([0, 80000]);
     
   
     const yFormat = d3.scaleLinear()
       .domain([0, d3.max(data, d => d[type])])
-      .range([0, 10]);
+      .range([0, 90000]);
 
     svg.append('g')
       .attr('class', 'axis-y-line')
-      .call(d3.axisLeft(y)
-        .tickFormat(d => `${yFormat(d)}%`));
+      .call(d3.axisLeft(y));
 
     const valueline = d3.line()
       .x(d => x(d.year))
@@ -129,21 +128,21 @@ class LineChart extends Component {
     svg.append('text')
       .attr('class', 'text-2015')
       .attr('x', x(1999))
-      .attr('y', y(2.75))
+      .attr('y', y(1200))
       .attr('font-size','20px')
-      .text('2.75');
+      .text('730');
 
     svg.append('text')
       .attr('class', 'text-2016')
       .attr('x', x(2002))
-      .attr('y', y(3))
+      .attr('y', y(1500))
       .attr('font-size','20px')
-      .text('3.25');
+      .text('1295');
 
     svg.append('circle')
       .attr('class', 'bubble-2015')
       .attr('cx', x(2002))
-      .attr('cy', y(2.75))
+      .attr('cy', y(3))
       .attr('r', 7)
       .style('fill', '#54EAEA')
       .style('opacity', 0.7);
@@ -151,19 +150,18 @@ class LineChart extends Component {
     svg.append('circle')
       .attr('class', 'bubble-2016')
       .attr('cx', x(1999))
-      .attr('cy', y(2.5))
+      .attr('cy', y(3))
       .attr('r', 7)
       .style('fill', '#54EAEA')
       .style('opacity', 0.7); 
     
 
-    const availableYears = [1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012,2013,2014,2015, 2016, 2017, 2018, 2019];
+      const availableYears = [1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019,2020,2021];
 
     // Add the X Axis
     svg.append('g')
       .attr('class', 'axis-x-line')
       .attr('transform', `translate(0,${innerHeight})`)
-      
       .call(d3.axisBottom(x)
         .tickValues(availableYears)
         .tickFormat(d3.format('.4')));
@@ -242,6 +240,8 @@ class LineChart extends Component {
       .attr('d', line.defined(d => d.defined));
 
     
+
+    
     this.userDataLine.forEach((data) => {
       console.log(`Year: ${data.year}, Y Value: ${data[type]}`);
     });
@@ -250,11 +250,12 @@ class LineChart extends Component {
     .data([this.userDataLine])
     .attr('d', line.defined(d => d.defined));
 
-
     const svg = d3.select('svg');
 
     const latestData = this.userDataLine[this.userDataLine.length - 1];
     const text = svg.selectAll('.value-text').data([latestData]);
+
+    
 
     text.exit().remove();
 
@@ -264,12 +265,12 @@ class LineChart extends Component {
       .attr('x', d => x(d.year))
       .attr('y', d => y(d[type]))
       .text(d => `${d[type]}`);
-  
 
+    if (!this.clipAnimation && d3.mean(this.userDataLine, d => d.defined) === 1) {
+        this.clipAnimation = true;
+        this.clipElement.transition().duration(1000).attr('width', x(dataXMax));
+      }
     
-    if (this.clipAnimation) {
-      this.userDataLine = this.userDataLine.filter(d => d.year <= year);
-    }
   });
 
   clampFunc = (a, b, c) => Math.max(a, Math.min(b, c));
@@ -285,14 +286,19 @@ class LineChart extends Component {
       })
       .filter(d => d.year >= startYear);
   }
-  
+
+  runAnimation = () => {
+    this.clipElement.transition().duration(1000).attr('width', this.props.dataXMax);
+  }
 
 
   render() {
     return (
       <div id="line-chart">
         <svg ref={this.svgReal} />
-         
+        { this.clipAnimation &&
+          <button onClick={this.runAnimation}>Run Animation</button>
+        }
       </div>
     );
   }
