+++ client/src/pages/survey/visualizations/LineChartVisual_DrugOverdose.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import * as d3 from 'd3';
+import { Divider, Typography, Container, Button } from "@mui/material";
 
 
 
@@ -27,6 +28,9 @@ class LineChart extends Component {
     this.youDrawIt = null;
     this.clipElement = null;
     this.clipAnimation = false;
+    this.state = {
+        showText: false
+      };
   }
 
   componentDidMount() {
@@ -40,12 +44,12 @@ class LineChart extends Component {
     this.userDataLine = this.transformData();
     const userDrawnValue = this.userDataLine.filter(d => d.defined === true);
 
-    const width = 500;
-    const height = 500;
+    const width = 900;
+    const height = 425;
     const margin = {
-      top: 100,
+      top: 50,
       right: 150,
-      bottom: 50,
+      bottom: 30,
       left: 150,
     };
     const svgContainer = d3.select(this.svgReal.current);
@@ -62,18 +66,18 @@ class LineChart extends Component {
 
    
     const x = d3.scaleLinear().range([0, innerWidth]);
+    this.setState({ x });
     const y = d3.scaleLinear().range([innerHeight, 0]);
-    y.domain([0, d3.max(data, d => d[type])]);
+    y.domain([0, 80000]);
     
   
     const yFormat = d3.scaleLinear()
       .domain([0, d3.max(data, d => d[type])])
-      .range([0, 10]);
+      .range([0, 90000]);
 
     svg.append('g')
-      .attr('class', 'axis-y-line')
-      .call(d3.axisLeft(y)
-        .tickFormat(d => `${yFormat(d)}%`));
+        .attr('class', 'axis-y-line')
+        .call(d3.axisLeft(y));
 
     const valueline = d3.line()
       .x(d => x(d.year))
@@ -129,21 +133,21 @@ class LineChart extends Component {
     svg.append('text')
       .attr('class', 'text-2015')
       .attr('x', x(1999))
-      .attr('y', y(2.75))
+      .attr('y', y(18849))
       .attr('font-size','20px')
-      .text('2.75');
+      .text('16849');
 
     svg.append('text')
       .attr('class', 'text-2016')
       .attr('x', x(2002))
-      .attr('y', y(3))
+      .attr('y', y(25849))
       .attr('font-size','20px')
-      .text('3.25');
+      .text('23518');
 
     svg.append('circle')
       .attr('class', 'bubble-2015')
       .attr('cx', x(2002))
-      .attr('cy', y(2.75))
+      .attr('cy', y(23518))
       .attr('r', 7)
       .style('fill', '#54EAEA')
       .style('opacity', 0.7);
@@ -151,7 +155,7 @@ class LineChart extends Component {
     svg.append('circle')
       .attr('class', 'bubble-2016')
       .attr('cx', x(1999))
-      .attr('cy', y(2.5))
+      .attr('cy', y(16849))
       .attr('r', 7)
       .style('fill', '#54EAEA')
       .style('opacity', 0.7); 
@@ -221,6 +225,15 @@ class LineChart extends Component {
       .call(this.mouseDragLine(x, y, dataXMax, valueline));
   };
 
+  renderAnimation = () => {
+    if (d3.mean(this.userDataLine, d => d.defined) === 1) {
+      const { data, type, startYear } = this.props;
+  
+      const dataXMax = d3.max(data, d => d.year);
+      this.clipElement.transition().duration(1000).attr('width', this.state.x(dataXMax));
+    }
+  };
+
 
 
   mouseDragLine = (x, y, dataXMax, line) => d3.drag().on('drag', () => {
@@ -250,7 +263,6 @@ class LineChart extends Component {
     .data([this.userDataLine])
     .attr('d', line.defined(d => d.defined));
 
-
     const svg = d3.select('svg');
 
     const latestData = this.userDataLine[this.userDataLine.length - 1];
@@ -267,9 +279,10 @@ class LineChart extends Component {
   
 
     
-    if (this.clipAnimation) {
-      this.userDataLine = this.userDataLine.filter(d => d.year <= year);
-    }
+    // if (!this.clipAnimation && d3.mean(this.userDataLine, d => d.defined) === 1) {
+    //     this.clipAnimation = true;
+    //     this.clipElement.transition().duration(1000).attr('width', x(dataXMax));
+    //   }
   });
 
   clampFunc = (a, b, c) => Math.max(a, Math.min(b, c));
@@ -285,16 +298,38 @@ class LineChart extends Component {
       })
       .filter(d => d.year >= startYear);
   }
-  
+
+  handleClick = () => {
+    this.setState({ showText: true });
+    this.renderAnimation();
+  };
 
 
   render() {
     return (
-      <div id="line-chart">
-        <svg ref={this.svgReal} />
-         
-      </div>
-    );
+        <div>
+          <div id="line-chart">
+            <svg ref={this.svgReal} />
+          </div>
+          <div style={{marginTop: '20px'}}>
+            <Button
+              variant="contained"
+              color="primary"
+              onClick={this.handleClick}
+              style={{marginTop: '120px',marginLeft: '500px', marginRight: '20px'}}
+            >
+              Complete
+            </Button>
+          </div>
+          { this.state.showText && (
+            <p align="justify">    has increased by more than <span style={{ fontWeight: "bold" }}>222.16 percent</span>.  In 2015, more Americans died from drug overdoses than from car accidents 
+            and gun homicides combined. It’s the worst drug overdose epidemic in American history, spurred by rising drug abuse, 
+            increased availability of prescription opioids and an influx of Data Sharing <span></span>potent synthetics like fentanyl and carfentanil. 
+            Drug overdoses are now the leading cause of death for Americans under 50.
+            </p>
+          ) }
+        </div>
+      );
   }
 }
 export default LineChart;
