+++ client/src/pages/survey/visualizations/LineChartVisual_Opioids.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import * as d3 from 'd3';
+import { Divider, Typography, Container, Button } from "@mui/material";
+
 
 
 
@@ -27,6 +29,9 @@ class LineChart extends Component {
     this.youDrawIt = null;
     this.clipElement = null;
     this.clipAnimation = false;
+    this.state = {
+        showText: false
+      };
   }
 
   componentDidMount() {
@@ -40,12 +45,12 @@ class LineChart extends Component {
     this.userDataLine = this.transformData();
     const userDrawnValue = this.userDataLine.filter(d => d.defined === true);
 
-    const width = 500;
-    const height = 500;
+    const width = 900;
+    const height = 425;
     const margin = {
-      top: 100,
+      top: 50,
       right: 150,
-      bottom: 50,
+      bottom: 100,
       left: 150,
     };
     const svgContainer = d3.select(this.svgReal.current);
@@ -62,18 +67,18 @@ class LineChart extends Component {
 
    
     const x = d3.scaleLinear().range([0, innerWidth]);
+    this.setState({ x });
     const y = d3.scaleLinear().range([innerHeight, 0]);
-    y.domain([0, d3.max(data, d => d[type])]);
+    y.domain([0, 80000]);
     
   
     const yFormat = d3.scaleLinear()
       .domain([0, d3.max(data, d => d[type])])
-      .range([0, 10]);
+      .range([0, 90000]);
 
     svg.append('g')
       .attr('class', 'axis-y-line')
-      .call(d3.axisLeft(y)
-        .tickFormat(d => `${yFormat(d)}%`));
+      .call(d3.axisLeft(y));
 
     const valueline = d3.line()
       .x(d => x(d.year))
@@ -129,21 +134,21 @@ class LineChart extends Component {
     svg.append('text')
       .attr('class', 'text-2015')
       .attr('x', x(1999))
-      .attr('y', y(2.75))
+      .attr('y', y(1200))
       .attr('font-size','20px')
-      .text('2.75');
+      .text('730');
 
     svg.append('text')
       .attr('class', 'text-2016')
       .attr('x', x(2002))
-      .attr('y', y(3))
+      .attr('y', y(1500))
       .attr('font-size','20px')
-      .text('3.25');
+      .text('1295');
 
     svg.append('circle')
       .attr('class', 'bubble-2015')
       .attr('cx', x(2002))
-      .attr('cy', y(2.75))
+      .attr('cy', y(3))
       .attr('r', 7)
       .style('fill', '#54EAEA')
       .style('opacity', 0.7);
@@ -151,19 +156,18 @@ class LineChart extends Component {
     svg.append('circle')
       .attr('class', 'bubble-2016')
       .attr('cx', x(1999))
-      .attr('cy', y(2.5))
+      .attr('cy', y(3))
       .attr('r', 7)
       .style('fill', '#54EAEA')
       .style('opacity', 0.7); 
     
 
-    const availableYears = [1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012,2013,2014,2015, 2016, 2017, 2018, 2019];
+      const availableYears = [1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019,2020,2021];
 
     // Add the X Axis
     svg.append('g')
       .attr('class', 'axis-x-line')
       .attr('transform', `translate(0,${innerHeight})`)
-      
       .call(d3.axisBottom(x)
         .tickValues(availableYears)
         .tickFormat(d3.format('.4')));
@@ -221,6 +225,15 @@ class LineChart extends Component {
       .call(this.mouseDragLine(x, y, dataXMax, valueline));
   };
 
+  renderAnimation = () => {
+    if (d3.mean(this.userDataLine, d => d.defined) === 1) {
+      const { data, type, startYear } = this.props;
+  
+      const dataXMax = d3.max(data, d => d.year);
+      this.clipElement.transition().duration(1000).attr('width', this.state.x(dataXMax));
+    }
+  };
+
 
 
   mouseDragLine = (x, y, dataXMax, line) => d3.drag().on('drag', () => {
@@ -242,6 +255,8 @@ class LineChart extends Component {
       .attr('d', line.defined(d => d.defined));
 
     
+
+    
     this.userDataLine.forEach((data) => {
       console.log(`Year: ${data.year}, Y Value: ${data[type]}`);
     });
@@ -250,11 +265,12 @@ class LineChart extends Component {
     .data([this.userDataLine])
     .attr('d', line.defined(d => d.defined));
 
-
     const svg = d3.select('svg');
 
     const latestData = this.userDataLine[this.userDataLine.length - 1];
     const text = svg.selectAll('.value-text').data([latestData]);
+
+    
 
     text.exit().remove();
 
@@ -264,12 +280,12 @@ class LineChart extends Component {
       .attr('x', d => x(d.year))
       .attr('y', d => y(d[type]))
       .text(d => `${d[type]}`);
-  
 
-    
-    if (this.clipAnimation) {
-      this.userDataLine = this.userDataLine.filter(d => d.year <= year);
-    }
+      // if (!this.clipAnimation && d3.mean(this.userDataLine, d => d.defined) === 1) {
+      //   this.clipAnimation = true;
+      //   this.clipElement.transition().duration(1000).attr('width', x(dataXMax));
+      // }
+
   });
 
   clampFunc = (a, b, c) => Math.max(a, Math.min(b, c));
@@ -285,16 +301,37 @@ class LineChart extends Component {
       })
       .filter(d => d.year >= startYear);
   }
-  
+
+  handleClick = () => {
+    this.setState({ showText: true });
+    this.renderAnimation();
+  };
 
 
   render() {
     return (
-      <div id="line-chart">
-        <svg ref={this.svgReal} />
-         
-      </div>
-    );
+        <div>
+          <div id="line-chart">
+            <svg ref={this.svgReal} />
+          </div>
+          <div style={{marginTop: '20px'}}>
+            <Button
+              variant="contained"
+              color="primary"
+              onClick={this.handleClick}
+              style={{marginTop: '60px',marginLeft: '500px', marginRight: '20px'}}
+            >
+              Complete
+            </Button>
+          </div>
+          { this.state.showText && (
+            <p align="justify">...has increased by more than <span style={{ fontWeight: "bold" }}>5451 percent</span>.  Substance use disorders refers to direct deaths from 
+            overdoses of <span style={{ fontWeight: "bold" }}>illicit drugs synthetic opioids (mostly fentanyl)</span> .  “We know that substance use is more dangerous than it has ever been, as fentanyl has continued to permeate the illicit drug supply, increasing the risk for overdoses among both people with substance use disorders as well as those who use drugs occasionally,” said Dr. Nora Volkow, director of the National Institute on Drug Abuse. Deaths involving synthetic opioids such as fentanyl increased by a marked 18% in 2021, according to the CDC data. Deaths involving cocaine and psychostimulants such as methamphetamine were also significantly more frequent, while those involving heroin decreased.
+            
+        </p>
+          ) }
+        </div>
+      );
   }
 }
 export default LineChart;
