+++ client/src/pages/survey/visualizations/LineChartVisual_Population.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import * as d3 from 'd3';
+import { Divider, Typography, Container, Button } from "@mui/material";
+
 
 
 
@@ -27,6 +29,9 @@ class LineChart extends Component {
     this.youDrawIt = null;
     this.clipElement = null;
     this.clipAnimation = false;
+    this.state = {
+        showText: false
+      };
   }
 
   componentDidMount() {
@@ -40,12 +45,12 @@ class LineChart extends Component {
     this.userDataLine = this.transformData();
     const userDrawnValue = this.userDataLine.filter(d => d.defined === true);
 
-    const width = 500;
-    const height = 500;
+    const width = 900;
+    const height = 425;
     const margin = {
-      top: 100,
+      top: 50,
       right: 150,
-      bottom: 50,
+      bottom: 30,
       left: 150,
     };
     const svgContainer = d3.select(this.svgReal.current);
@@ -62,6 +67,7 @@ class LineChart extends Component {
 
    
     const x = d3.scaleLinear().range([0, innerWidth]);
+    this.setState({ x });
     const y = d3.scaleLinear().range([innerHeight, 0]);
     y.domain([0, d3.max(data, d => d[type])]);
     
@@ -220,6 +226,14 @@ class LineChart extends Component {
       .attr('height', innerHeight)
       .call(this.mouseDragLine(x, y, dataXMax, valueline));
   };
+  renderAnimation = () => {
+    if (d3.mean(this.userDataLine, d => d.defined) === 1) {
+      const { data, type, startYear } = this.props;
+  
+      const dataXMax = d3.max(data, d => d.year);
+      this.clipElement.transition().duration(1000).attr('width', this.state.x(dataXMax));
+    }
+  };
 
 
 
@@ -250,6 +264,10 @@ class LineChart extends Component {
     .data([this.userDataLine])
     .attr('d', line.defined(d => d.defined));
 
+    // if (!this.clipAnimation && d3.mean(this.userDataLine, d => d.defined) === 1) {
+    //   this.clipAnimation = true;
+    //   this.clipElement.transition().duration(1000).attr('width', x(dataXMax));
+    // }
 
     const svg = d3.select('svg');
 
@@ -285,16 +303,39 @@ class LineChart extends Component {
       })
       .filter(d => d.year >= startYear);
   }
+  handleClick = () => {
+    this.setState({ showText: true });
+    this.renderAnimation();
+  };
   
 
 
   render() {
     return (
-      <div id="line-chart">
-        <svg ref={this.svgReal} />
-         
-      </div>
-    );
+        <div>
+          <div id="line-chart">
+            <svg ref={this.svgReal} />
+          </div>
+          <div style={{marginTop: '20px'}}>
+            <Button
+              variant="contained"
+              color="primary"
+              onClick={this.handleClick}
+              style={{marginTop: '120px',marginLeft: '500px', marginRight: '20px'}}
+            >
+              Complete
+            </Button>
+          </div>
+          { this.state.showText && (
+            <p align="justify">...has increased by more than <span style={{ fontWeight: "bold" }}>137 percent</span>.  The United States is currently in the grips of a powerful drug epidemic, 
+            with the share of population with drug use disorders steadily climbing every year. A drug use disorder is a mental disorder that affects a person’s brain and behavior, leading to a person’s 
+            inability to control their use of drugs including legal or illegal drugs. Drug use disorders occur when an individual 
+            compulsively misuses drugs or alcohol and continues abusing the substance despite knowing the negative impact it has on their life.
+            
+            </p>
+          ) }
+        </div>
+      );
   }
 }
 export default LineChart;
